Move script files to scripts directory.
