+++ scripts/updateApis.ts
@@ -7,13 +7,13 @@
 
 import path from 'path';
 import fs from 'fs-extra';
-import {updateApis} from './scripts/utils';
-import {ASSET_OBJECT_MAP} from './scripts/config';
+import {updateApis} from './utils';
+import {ASSET_OBJECT_MAP} from './config';
 
 const API_NAMES = Object.keys(ASSET_OBJECT_MAP);
 
-const OLD_APIS_PATH = path.join(__dirname, 'temp/oldApis');
-const PRODUCTION_API_PATH = path.join(__dirname, 'apis');
+const OLD_APIS_PATH = path.join(__dirname, '../temp/oldApis');
+const PRODUCTION_API_PATH = path.join(__dirname, '../apis');
 
 // DOWNLOAD PRODUCTION DATA
 fs.moveSync(PRODUCTION_API_PATH, OLD_APIS_PATH, {overwrite: true});
